+++ multipicker/src/main/java/im/vector/lib/multipicker/MultiPicker.kt
@@ -25,6 +25,7 @@ class MultiPicker<T> {
         val AUDIO by lazy { MultiPicker<AudioPicker>() }
         val CONTACT by lazy { MultiPicker<ContactPicker>() }
         val CAMERA by lazy { MultiPicker<CameraPicker>() }
+        val RECVIDEO by lazy { MultiPicker<RecordVideoPicker>() }
 
         @Suppress("UNCHECKED_CAST")
         fun <T> get(type: MultiPicker<T>): T {
@@ -35,6 +36,7 @@ class MultiPicker<T> {
                 AUDIO   -> AudioPicker() as T
                 CONTACT -> ContactPicker() as T
                 CAMERA  -> CameraPicker() as T
+                RECVIDEO -> RecordVideoPicker() as T
                 else    -> throw IllegalArgumentException("Unsupported type $type")
             }
         }